Move location.py and spacetime.py to models
