+++ src/virtualship/instruments/ctd.py
@@ -7,7 +7,7 @@ from pathlib import Path
 import numpy as np
 from parcels import FieldSet, JITParticle, ParticleSet, Variable
 
-from ..spacetime import Spacetime
+from ..models.spacetime import Spacetime
 
 
 @dataclass
